Finalizing approval button

+++ Jenkinsfile
@@ -36,7 +36,9 @@ pipeline {
                     export PATH=$PATH:/home/jenkins/ovftool
 
                     echo "Remove the current terraform state"
+                    if [ -f "terraform.tfstate" ]; then
                     rm terraform.tfstate
+                    fi
 
                     terraform init
                     terraform plan -input=false -out tfplan
@@ -44,10 +46,6 @@ pipeline {
                 
                 sh 'terraform show tfplan'
 
-                // Create an Approval Button with a timeout of 15minutes.
-                timeout(time: 15, unit: "MINUTES") {
-                    input message: 'Do you want to approve the deployment?', ok: 'Yes'
-                }
             }
         }
 
@@ -59,6 +57,10 @@ pipeline {
             }
 
             steps{
+                // Create an Approval Button with a timeout of 15minutes.
+                timeout(time: 15, unit: "MINUTES") {
+                    input message: 'Do you want to approve the deployment? Look at the tfplan that was created on the previous stage.', ok: 'Yes'
+                }
                 echo 'Provisioning the VMs' 
                 sh '''
                     #!/bin/bash
